+++ earthdaily/earthone/dynamic_compute/image_stack.py
@@ -177,6 +177,7 @@ class ImageStackSerializationModel(BaseSerializationModel):
     predicate_filter: Optional[str] = None
     sort_by: Optional[str] = None
     ascending: Optional[bool] = None
+    obj_type: Optional[str] = None
 
     @classmethod
     def from_json(cls, data: str) -> ImageStackSerializationModel:
@@ -233,6 +234,7 @@ class ImageStack(
         predicate_filter: Optional[str] = None,
         sort_by: Optional[str] = None,
         ascending: Optional[bool] = None,
+        obj_type: Optional[str] = None,
     ):
         """
         Initialize a new instance of ImageStack. Users should rely on
@@ -276,6 +278,9 @@ class ImageStack(
             "pad": pad,
             "resampler": resampler,
         }
+
+        if obj_type is not None and obj_type != "ImageStack":
+            raise ValueError(f"Object {obj_type} is not an ImageStack")
 
     @classmethod
     def from_product_bands(
@@ -676,6 +681,7 @@ class ImageStack(
             bands=self.bands,
             start_datetime=self.start_datetime,
             end_datetime=self.end_datetime,
+            obj_type="ImageStack",
         ).json()
 
     @classmethod